Initial prototype of encrypted secure.config
This initial version is a pure copy of the Gerrit DefaultSecureStore,
which contains the basic structure of the plugin and the associated
documentation.

Main purpose of this commit is to demonstrate that we can build
the plugin, install on Gerrit /lib, configure and have it fully
working with a secure.config file.

Change-Id: I60e4754a7fc2826362784c25ca59305c43015ff5